Change arguments of functions related to abbreviating class-tags
Class-tags (`eieio-class-tag--CLASS') can be abbreviated for storage
in the database, to allow simpler SQL queries.  It is even possible
to abbreviate the `CLASS' part, e.g. if classes have the form
`PREFIX-INFIX[-SUFFIX]' then that can be abbreviated to just `INFIX'.

This is done by the method `closql--class-to-sql' and the reverse is
done by `closql--sql-to-class'.  Until now users of this library had
to implement those methods for their database object.  But that has
two issues:

* Each user has to implement those methods again.  The next commit
  will make that unnecessary.

* The methods are associated with an object of the database class.
  This makes it impossible to use this feature when the database
  contains objects of different types in different tables.

Fix the second issue here by associating these methods with the class
of the base object(s) instead.  Some callers are adjusted accordingly.
Among those `closql-where-class-in' gets a new function signature due
to this.