closql-{entries,select}: rename CLASSES argument to PRED
While this can be a list of classes (or predicates) it usually
is a single class predicate (less often a list of predicates.
A list of classes is the least likely value, so we probably
should not name the argument after that.

Also a new CLASS argument will be added in the next commit, and
having a CLASS and a CLASSES argument would be confusing.