closql--db-init: remove method for closql-database
Now users of this library have to implement this themselves.  They
have to define a `closql--db-init' method for their database class.

The recommended way of doing that is to define the table schemata
explicitly instead of generating them from the database and object
class definitions as was done here before.  As the old approach has
proven to not be to rigid.  Theoretically it would be possible to
make the old approach more flexible, but that would result in a lot
of complexity and the need for later adjustments, so it is better
to embrace a tiny bit of duplication.