Allow storing a list of objects in an object slot
The value of a slot can now be a list of objects.  Like for a slot
whose value is a list of lists, the elements have to be uniform, in
this case they have to share a base-class, and they are unordered.

Objects in such secondary tables can be queried directly, using the
same functions that are used to query objects in the primary table.
For secondary tables the new optional ~class~ argument of these query
functions has to be provided.