Make sure to only call eieio--class-name on a class
In Emacs 25 an earlier eieio objects were represented using vectors
and the type descriptor was a symbol.

Since Emacs 26 eieio objects use a class record (with circular
references) as the type descriptor of the object record.  Records
did not exist before that release, so testing whether the function
`record' exists was a bit an indirect but appropriate in method to
check what kind of type descriptor we have to deal with.

Since Emacs 27 symbols are used as type descriptor again (and records
are still being used for eieio objects).  So we now have to check the
type of the type descriptor itself (symbol or `eieio--class' record).

Also see https://github.com/emacs-mirror/emacs/commit/5f01af6c8e0f73.

Fixes https://github.com/magit/forge/issues/159.