eieio-oref--closql-oref: Prevent recursive load
When a class has been autoloaded and the resulting dummy class object
is encountered by `object-of-class-p' (or a `*--eieio-childp' wrapper),
then that calls `eieio--full-class-object', which then loads the
library that fully defines the class.

If this happens while that library is being loaded, then that causes
a recursive load because we have adviced `eieio-oref' to do a type
check.

Prevent this by using `eieio--class-object' instead of
`eieio--full-class-object'.  The latter does not load any libraries.