Adjust to how unbound slots are represented in Emacs 28
In Eieio unbound slots are represented using an uninterned symbol,
which is stored as the value of a variable.

The name of the symbol used to be `unbound' and the variable was named
`eieio-unbound'.  Since 4c6554413d318069239ba83f4f42fa2452801d30 both
the uninterned symbol and the variable are named `eieio--unbound'.

We adjust to that by using the variable `eieio--unbound' going forward
and defining it as a variable for `eieio-unbound' for older Emacsen.

Unlike Eieio itself we never used `unbound' as the representation of
an unbound slot when saving objects in the database, because that is
likely to appear as an actual value.  Instead we used `eieio-unbound'.

We could change that to `eieio--unbound' but don't do so because the
risk of complications down the line are greater than the benefit of a
cosmetic increase in consistency.