Support for alternative connectors
The `closql-database' class is abstract now, making it necessary to
define a class that derives from that and also from a connector class.
The choice of which connector to use can be left to the user.  It may
be necessary to define a `closql-db' method for alternative connectors
and possibly other methods as well.

  (cl-case forge-database-connector
    (sqlite
     (defclass forge-database (emacsql-sqlite-connection closql-database)
       ((object-class :initform 'forge-repository))))
    (libsqlite3
     (require (quote emacsql-libsqlite3))
     (with-no-warnings
       (defclass forge-database (emacsql-libsqlite3-connection closql-database)
         ((object-class :initform 'forge-repository))))))