Database classes no longer derive from a connector class
Back when this library supported only a single connector class, it
made sense that a concrete database class had to derive from both
that connector class and `closql-database'.

Now that multiple SQLite connectors exist but their availability
depends on the Emacs version in use and built time settings, packages
should automatically use the best connector.  If the package's
database class has to derive from a connector class, then that is
painful to do.

So we add some indirection.  A package's database class only has to
derive from `closql-database' and the connector object is stored in
the new `connector' slot of the database object.

Implement some generic functions for `closql-database' to make certain
downstream changes unnecessary.  `emacsql-with-transaction' is a macro
so we cannot do the same for that, so provide `closql-with-transaction'
as a substitute.